Factored out test code from loader.rs
